requirements.txt
alembic==1.16.4
appdirs==1.4.4
BareNecessities==0.2.8
bcrypt==4.3.0
black==25.1.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.2.1
colorama==0.4.6
comtypes==1.4.10
cssselect==1.2.0
demoji==1.1.0
dnspython==2.7.0
email_validator==2.2.0
emoji==2.14.1
et_xmlfile==2.0.0
Faker==37.5.3
Flask==3.1.1
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
gTTS==2.5.4
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
lxml_html_clean==0.4.1
Mail==2.1.0
Mako==1.3.10
MarkupSafe==3.0.2
mutagen==1.47.0
mypy_extensions==1.1.0
numpy==2.3.2
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
parse==1.20.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.3.8
pyee==11.1.1
pypiwin32==223
pyppeteer==2.0.0
pyquery==2.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.98
pytz==2025.2
pywin32==308
requests==2.32.3
requests-html==0.10.0
six==1.17.0
soupsieve==2.6
SQLAlchemy==2.0.43
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
w3lib==2.3.1
websockets==10.4
Werkzeug==3.1.3
WTForms==3.2.1
WTForms-SQLAlchemy==0.4.2
zipp==3.21.0
